Reverts the use of "corejs" in Babel runtime transform
Because it introduced an unexpected dependency on
`@babel/runtime-corejs3`, which is not co-installed automatically with
`@babel/runtime`, and in general is not clearly mentioned in Babel docs.
Thus, it will be better to keep CoreJS polyfills out of the library
build, and let to the end host code to provide CoreJS polyfills when
needed.